users_service/pyproject.toml
[project]
name = "users_services"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "bcrypt>=5.0.0",
    "cryptography>=46.0.3",
    "fastapi[standard]>=0.128.0",
    "grpcio>=1.76.0",
    "grpcio-tools>=1.76.0",
    "httpx>=0.28.1",
    "isort>=7.0.0",
    "mypy>=1.19.1",
    "pydantic-settings>=2.12.0",
    "pyjwt>=2.10.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
]
